// src/libs/api/schema/core-services.d.ts
export interface components {
    schemas: {
        /** Address */
        Address: {
            /** Pk */
            pk: number;
            address_type: components["schemas"]["AddressType"];
            /** Name */
            name: string | null;
            /** Email */
            email: string | null;
            /** Address */
            address: string;
            /** Phone */
            phone: string | null;
            /**
             * Created At
             * Format: date-time
             */
            created_at: string;
            /**
             * Last Modified At
             * Format: date-time
             */
            last_modified_at: string;
            created_by: components["schemas"]["UserMin"];
            last_modified_by: components["schemas"]["UserMin"] | null;
            contact: components["schemas"]["ContactMin"];
        };
// ...
        Contact: {
            /** Pk */
            pk: number;
            /** Name */
            name: string;
            legal_type: components["schemas"]["ContactLegalType"];
            /** Tax Id */
            tax_id: string | null;
            image: components["schemas"]["ContactImageURL"];
            /**
             * Created At
             * Format: date-time
             */
            created_at: string;
            /**
             * Last Modified At
             * Format: date-time
             */
            last_modified_at: string;
            created_by: components["schemas"]["UserMin"];
            last_modified_by: components["schemas"]["UserMin"] | null;
            /** Deleted At */
            deleted_at: string | null;
            parent: components["schemas"]["ContactMin"] | null;
            /** Tags */
            tags: components["schemas"]["TagMin"][];
        };
// ...
        Industry: {
            /** Pk */
            pk: number;
            industry_type: components["schemas"]["IndustryType"];
            /**
             * Created At
             * Format: date-time
             */
            created_at: string;
            /**
             * Last Modified At
             * Format: date-time
             */
            last_modified_at: string;
            created_by: components["schemas"]["UserMin"];
            last_modified_by: components["schemas"]["UserMin"] | null;
            contact: components["schemas"]["ContactMin"];
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}